+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     * Add update method to update OSM
   * Osm::FlexiRecord
     * Add add\_field method to add a field to the record in OSM
+    * Add update\_field method to rename a field in OSM
 
 ## Version 0.1.11